@@ -2,18 +2,18 @@
 name = "distance-db-populator"
 version = "0.1.0"
 authors = ["Brian Bowman <seeker14491@gmail.com>"]
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 anyhow = "1"
 az = "1"
-color-backtrace = "0.6"
+color-backtrace = "0.7"
 distance-steam-data-client = { git = "https://github.com/Seeker14491/DistanceSteamDataServer.git" }
 distance-util = { git = "https://github.com/Seeker14491/distance-util.git", tag = "v0.3.0" }
 dotenv = "0.15"
 futures = "0.3"
 indicatif = "0.17"
-itertools = "0.13"
+itertools = "0.14"
 num-traits = "0.2"
 reqwest = { version = "0.12", default-features = false, features = [
   "gzip",
